[FEATURE] 노트 도메인의 서비스 단위 테스트(진행 중)

See merge request s09-final/S09P31A407!48

+++ backend/src/main/java/com/ziglog/ziglog/domain/note/service/NoteServiceImpl.java
@@ -26,10 +26,7 @@ public class NoteServiceImpl implements NoteService{
 
     //Note
     @Override
-    public Note createNote() {
-
-        Member member = new Member();//SecurityContext 내의 사용자로 변할 것
-
+    public Note createNote(Member member) {
         Note note = Note.builder()
                     .author(member)
                     .build();
@@ -37,12 +34,11 @@ public class NoteServiceImpl implements NoteService{
     }
 
     @Override
-    public Note saveNoteWithDiff(Note note) throws Exception {
-
-        //글의 저자가 Security Context 내의 유저와 같은지 확인
+    public Note modifyNote(Member member, Note note) throws Exception {
+        if (!checkOwner(member, note)) throw new Exception();
 
         //영속성 Context 내의 노트
-        Note origin = noteRepository.findNoteById(note.getId()).orElseThrow(() -> new Exception());
+        Note origin = noteRepository.findNoteById(note.getId()).orElseThrow(Exception::new);
         origin.setTitle(note.getTitle());//타이틀
         origin.setContent(note.getContent());//컨텐츠
         origin.setBrief(note.getBrief());//목록 프리뷰
@@ -52,9 +48,10 @@ public class NoteServiceImpl implements NoteService{
     }
 
     @Override
-    public Note setPublic(Note note) throws Exception {
-        Note origin = noteRepository.findNoteById(note.getId()).orElseThrow(() -> new Exception());
+    public Note setPublic(Member member, Note note) throws Exception {
+        if (!checkOwner(member, note)) throw new Exception();
 
+        Note origin = noteRepository.findNoteById(note.getId()).orElseThrow(Exception::new);
         origin.setPublic(note.isPublic());
         if (origin.isPublic() && origin.getPostDatetime() == null) {
             origin.setPostDatetime(LocalDateTime.now());
@@ -64,7 +61,9 @@ public class NoteServiceImpl implements NoteService{
     }
 
     @Override
-    public Boolean deleteNote(Long noteId) {
+    public Boolean deleteNote(Member member, Long noteId) throws Exception {
+        Note note = noteRepository.findNoteById(noteId).orElseThrow(Exception::new);
+        if (!checkOwner(member, note)) throw new Exception();
 
         //삭제 요청자가 Security Context 내의 사용자 같은지 확인
         try {
@@ -78,32 +77,38 @@ public class NoteServiceImpl implements NoteService{
 
     @Override
     public Note getNote(Long noteId) throws Exception{
-        return noteRepository.findNoteById(noteId).orElseThrow(() -> new Exception());
+        return noteRepository.findNoteById(noteId).orElseThrow(Exception::new);
     }
 
     @Override
-    public List<Note> findNotesQuotingThisNote(Long noteId) {
+    public List<Note> findNotesQuotingThisNote(Long noteId) throws Exception {
+        //TODO
+        Note note = noteRepository.findNoteById(noteId).orElseThrow(Exception::new);
         return null;
     }
 
     // Folder
     @Override
-    public Folder addFolder(Folder folder) {
+    public Folder addFolder(Member member, Folder folder) {
+        //TODO
         return folderRepository.save(folder);
     }
 
     @Override
-    public Folder modifyFolder(Folder folder) throws Exception {
-
+    public Folder modifyFolder(Member member, Folder folder) throws Exception {
         //JPA 영속성 컨테스트 내
-        Folder origin = folderRepository.findById(folder.getId()).orElseThrow(() -> new Exception());
+        if (!checkOwner(member, folder)) throw new Exception();
+
+        Folder origin = folderRepository.findById(folder.getId()).orElseThrow(Exception::new);
         origin.setTitle(folder.getTitle());
 
         return null;
     }
 
     @Override
-    public Boolean deleteFolder(Long folderId) {
+    public Boolean deleteFolder(Member member, Long folderId) throws Exception {
+        Folder folder= folderRepository.findById(folderId).orElseThrow(Exception::new);
+        if (!checkOwner(member, folder)) throw new Exception();
 
         try {
             folderRepository.deleteById(folderId);
@@ -111,21 +116,20 @@ public class NoteServiceImpl implements NoteService{
         catch (Exception e) {
             return false;
         }
-
         return true;
     }
 
     @Override
     public List<Folder> listFolder(String nickname) throws Exception {
-        Member user = memberRepository.findByNickname(nickname).orElseThrow(() -> new Exception());
+        Member user = memberRepository.findByNickname(nickname).orElseThrow(Exception::new);
         return user.getFolders();
     }
 
     // Quotation
     @Override
-    public Boolean addQuotation(Long fromNote, Long toNote) throws Exception{
-        Note from = noteRepository.findNoteById(fromNote).orElseThrow(() -> new Exception() );
-        Note to = noteRepository.findNoteById(toNote).orElseThrow(() -> new Exception());
+    public Boolean addQuotation(Member member, Long fromNote, Long toNote) throws Exception{
+        Note from = noteRepository.findNoteById(fromNote).orElseThrow(Exception::new);
+        Note to = noteRepository.findNoteById(toNote).orElseThrow(Exception::new);
 
         Quotation quotation = Quotation.builder()
                                 .startNote(from)
@@ -141,12 +145,22 @@ public class NoteServiceImpl implements NoteService{
     }
 
     @Override
-    public Boolean deleteQuotation(Long fromNote, Long toNote) throws Exception {
-        Note from = noteRepository.findNoteById(fromNote).orElseThrow(() -> new Exception() );
-        Note to = noteRepository.findNoteById(toNote).orElseThrow(() -> new Exception());
+    public Boolean deleteQuotation(Member member, Long fromNote, Long toNote) throws Exception {
+        Note from = noteRepository.findNoteById(fromNote).orElseThrow(Exception::new);
+        Note to = noteRepository.findNoteById(toNote).orElseThrow(Exception::new);
 
         Quotation quotation = quotationRepository.findByStartNoteAndEndNote(from, to);
         quotationRepository.deleteQuotationById(quotation.getId());
         return false;
+    }
+
+    @Override
+    public Boolean checkOwner(Member member, Note note){
+        return note.getAuthor() == member;
+    }
+
+    @Override
+    public Boolean checkOwner(Member member, Folder folder){
+        return folder.getOwner() == member;
     }
 }
